+++ __test__/src/main.rs
@@ -1,23 +1,9 @@
-use std::env::current_dir;
-
-use dotenv_plus::{
-    common::get_rust_env,
-    env::{DotEnv, Environment},
-    var::var,
-};
+use dotenv_plus::{DotEnv, var};
 
 fn main() {
-    let env: Environment = if cfg!(feature = "prd") {
-        Environment::Production
-    } else if cfg!(feature = "test") {
-        Environment::Test
-    } else {
-        Environment::Development
-    };
+    DotEnv::new().run();
 
-    DotEnv::new().dir(current_dir().unwrap()).env(env.as_code()).done();
-
-    println!("Rust environment: {}", get_rust_env());
+    println!("Rust environment: {}", var("RUST_ENV"));
     println!("A: {}", var("A"));
     println!("B: {}", var("B"));
     println!("C: {}", var("C"));
@@ -25,37 +11,67 @@ fn main() {
 
 #[cfg(test)]
 mod tests {
-    use std::env::current_dir;
-
-    use dotenv_plus::{
-        common::{
-            get_rust_env, is_dev, is_environment_development,
-            is_environment_production, is_environment_test, is_prd, is_test,
-        },
-        env::{DotEnv, Environment},
-        var::{get_var, var},
-    };
+    use dotenv_plus::{DotEnv, get_var, set_var, set_vars, var};
 
     #[test]
-    fn main() {
-        DotEnv::new()
-            .dir(current_dir().unwrap())
-            .environment(Environment::Test.as_code())
-            .done();
+    fn test_env() {
+        // development
 
-        assert_eq!(get_rust_env(), "test");
-        assert_eq!(is_environment_development(), false);
-        assert_eq!(is_dev(), false);
-        assert_eq!(is_environment_test(), true);
-        assert_eq!(is_test(), true);
-        assert_eq!(is_environment_production(), false);
-        assert_eq!(is_prd(), false);
+        DotEnv::new().run();
+
+        assert_eq!(var("RUST_ENV"), "development");
 
         assert_eq!(get_var("A").unwrap(), ".env");
         assert_eq!(var("A"), ".env");
+
+        assert_eq!(get_var("B").unwrap(), ".env.development");
+        assert_eq!(var("B"), ".env.development");
+
+        assert_eq!(get_var("C").unwrap(), ".env.development");
+        assert_eq!(var("C"), ".env.development");
+
+        // test
+
+        DotEnv::new().environment("test").run();
+
+        assert_eq!(var("RUST_ENV"), "test");
+
+        assert_eq!(get_var("A").unwrap(), ".env");
+        assert_eq!(var("A"), ".env");
+
         assert_eq!(get_var("B").unwrap(), ".env.test");
         assert_eq!(var("B"), ".env.test");
+
         assert_eq!(get_var("C").unwrap(), ".env.test");
         assert_eq!(var("C"), ".env.test");
+    }
+
+    #[test]
+    fn test_set_vars_get_var() {
+        set_vars(vec![
+            ("__TEST_SET_VARS_1", "abc1"),
+            ("__TEST_SET_VARS_2", "abc2"),
+        ]);
+
+        assert_eq!(get_var("__TEST_SET_VARS_1").unwrap(), "abc1");
+        assert_eq!(var("__TEST_SET_VARS_1"), "abc1");
+
+        assert_eq!(get_var("__TEST_SET_VARS_2").unwrap(), "abc2");
+        assert_eq!(var("__TEST_SET_VARS_2"), "abc2");
+    }
+
+    #[test]
+    fn test_set_var_get_var() {
+        set_var("__TEST_SET_VAR", "abc");
+
+        assert_eq!(get_var("__TEST_SET_VAR").unwrap(), "abc");
+        assert_eq!(var("__TEST_SET_VAR"), "abc");
+    }
+
+    #[test]
+    fn test_get_var_error() {
+        if let Ok(_) = get_var("__TEST_GET_VAR_ERROR") {
+            panic!();
+        }
     }
 }