+++ package/src/var.rs
@@ -3,12 +3,12 @@ use std::{
     ffi::OsStr,
 };
 
-/// Set multiple environment variable for the current process.
+/// Set multiple environment variables for the current process.
 ///
 /// ## Example
 ///
 /// ```no_run
-/// use dotenv_plus::var::set_vars;
+/// use dotenv_plus::set_vars;
 ///
 /// set_vars(vec![
 ///     ("key1", "value1"),
@@ -19,12 +19,12 @@ pub fn set_vars(vars: Vec<(&str, &str)>) {
     vars.iter().for_each(|(k, v)| set_var(k, v));
 }
 
-/// Set environment variable for the current process.
+/// Set an environment variable for the current process.
 ///
 /// ## Example
 ///
 /// ```no_run
-/// use dotenv_plus::var::set_var;
+/// use dotenv_plus::set_var;
 ///
 /// set_var("key", "value");
 /// ```
@@ -40,7 +40,7 @@ pub fn set_var<K: AsRef<OsStr>, V: AsRef<OsStr>>(
 /// ## Example
 ///
 /// ```no_run
-/// use dotenv_plus::var::get_vars;
+/// use dotenv_plus::get_vars;
 ///
 /// let vars: Vec<(String, String)> = get_vars();
 /// ```
@@ -48,32 +48,32 @@ pub fn get_vars() -> Vec<(String, String)> {
     vars().collect()
 }
 
-/// Get environment variable.
+/// Get an environment variable.
 ///
-/// Use [`var`] to handle the error automatically.
+/// Use [`var`] to handle errors automatically.
 ///
 /// ## Example
 ///
 /// ```no_run
-/// use dotenv_plus::var::get_var;
+/// use dotenv_plus::get_var;
 ///
-/// let environment: String = get_var("RUST_ENV").unwrap();
+/// let env: String = get_var("RUST_ENV").unwrap();
 /// ```
 pub fn get_var<K: AsRef<OsStr>>(name: K) -> Result<String, VarError> {
     _var(name)
 }
 
-/// Get environment variable as string,
-/// and panic if the environment variable is missing.
+/// Get an environment variable as string,
+/// and panic if it is missing.
 ///
-/// Use [`get_var`] to handle the error manually.
+/// Use [`get_var`] to handle errors manually.
 ///
 /// ## Example
 ///
 /// ```no_run
-/// use dotenv_plus::var::var;
+/// use dotenv_plus::var;
 ///
-/// let environment: String = var("RUST_ENV");
+/// let env: String = var("RUST_ENV");
 /// ```
 pub fn var<K: AsRef<OsStr>>(name: K) -> String {
     match _var(&name) {